feat(studio): Add Aura deposits
