+++ src/apps/aura/arbitrum/aura.deposit.token-fetcher.ts
@@ -1,0 +1,8 @@
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+
+import { AuraDepositTokenFetcher } from '../common/aura.deposit.token-fetcher';
+
+@PositionTemplate()
+export class ArbitrumAuraDepositTokenFetcher extends AuraDepositTokenFetcher {
+  boosterAddresses = ['0x98ef32edd24e2c92525e59afc4475c1242a30184'];
+}
