+++ src/apps/aura/aura.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 
 import { AbstractApp } from '~app/app.dynamic-module';
 
+import { ArbitrumAuraDepositTokenFetcher } from './arbitrum/aura.deposit.token-fetcher';
 import { AuraBalancerPoolResolver } from './common/aura.balancer-pool.resolver';
 import { AuraContractFactory } from './contracts';
 import { EthereumAuraAuraBalCompounderContractPositionFetcher } from './ethereum/aura.aura-bal-compounder.contract-position-fetcher';
@@ -12,12 +13,15 @@ import { EthereumAuraDepositTokenFetcher } from './ethereum/aura.deposit.token-f
 import { EthereumAuraLockerContractPositionFetcher } from './ethereum/aura.locker.contract-position-fetcher';
 import { EthereumAuraLpFarmContractPositionFetcher } from './ethereum/aura.lp-farm.contract-position-fetcher';
 import { EthereumAuraStakedAuraBalTokenFetcher } from './ethereum/aura.staked-aura-bal.token-fetcher';
+import { OptimismAuraDepositTokenFetcher } from './optimism/aura.deposit.token-fetcher';
 
 @Module({
   providers: [
     AuraContractFactory,
     // helpers
     AuraBalancerPoolResolver,
+    // Arbitrum
+    ArbitrumAuraDepositTokenFetcher,
     // Ethereum
     EthereumAuraAuraBalTokenFetcher,
     EthereumAuraDepositTokenFetcher,
@@ -27,6 +31,8 @@ import { EthereumAuraStakedAuraBalTokenFetcher } from './ethereum/aura.staked-au
     EthereumAuraAuraBalStakingContractPositionFetcher,
     EthereumAuraStakedAuraBalTokenFetcher,
     EthereumAuraAuraBalCompounderContractPositionFetcher,
+    // Optimism
+    OptimismAuraDepositTokenFetcher,
   ],
 })
 export class AuraAppModule extends AbstractApp() {}
